app-api/requirements.txt
beautifulsoup4==4.11.1
fastapi==0.88.0
pydantic==1.10.4
python-dateutil==2.8.2
requests==2.28.1
torch==1.12
transformers==4.25.1
uvicorn==0.20.0
